fix: 해지 후 금액 double로 변경

+++ hanaUp/src/After_DeleteSavings/DeleteSavings.jsx
@@ -155,7 +155,7 @@ const DeleteSavings = () => {
               해지 후 받으시는 금액 (세후)
             </div>
             <div style={{ ...font.header.h1, color: '#2d2d2d', width: '100%' }}>
-              {currencySymbol} {parseInt(Number(savings.finalAmount)).toLocaleString()}
+              {currencySymbol} {parseFloat(Number(savings.finalAmount)).toLocaleString()}
             </div>
           </TextMainContainer>
           <Horizon />
